app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
pypdf
python-dotenv